chore(user/api): move User to model
