+++ backend/src/main/java/com/techfun/altrua/features/auth/repository/RefreshTokenRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Lock;
 
 import com.techfun.altrua.features.auth.domain.RefreshToken;
-import com.techfun.altrua.features.user.domain.User;
+import com.techfun.altrua.features.user.domain.model.User;
 
 import jakarta.persistence.LockModeType;
