+++ backend/src/main/java/com/techfun/altrua/features/auth/service/AuthService.java
@@ -16,7 +16,7 @@ import com.techfun.altrua.features.auth.api.dto.AuthResponseDTO;
 import com.techfun.altrua.features.auth.api.dto.LoginRequestDTO;
 import com.techfun.altrua.features.auth.api.dto.RegisterUserRequestDTO;
 import com.techfun.altrua.features.auth.api.dto.RotateResult;
-import com.techfun.altrua.features.user.domain.User;
+import com.techfun.altrua.features.user.domain.model.User;
 import com.techfun.altrua.features.user.repository.UserRepository;
 import com.techfun.altrua.infra.security.jwt.JwtProvider;
 import com.techfun.altrua.infra.security.jwt.JwtValidator;
