+++ backend/src/main/java/com/techfun/altrua/features/event/domain/model/Event.java
@@ -13,7 +13,7 @@ import com.techfun.altrua.core.common.exceptions.DomainException;
 import com.techfun.altrua.features.event.domain.enums.EventStatusEnum;
 import com.techfun.altrua.features.ong.domain.model.Ong;
 import com.techfun.altrua.features.tag.domain.Tag;
-import com.techfun.altrua.features.user.domain.User;
+import com.techfun.altrua.features.user.domain.model.User;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
