+++ backend/src/main/java/com/techfun/altrua/features/user/domain/model/User.java
@@ -1,4 +1,4 @@
-package com.techfun.altrua.features.user.domain;
+package com.techfun.altrua.features.user.domain.model;
 
 import java.time.Instant;
 import java.util.UUID;
