+++ backend/src/main/java/com/techfun/altrua/features/user/repository/UserRepository.java
@@ -5,7 +5,7 @@ import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.techfun.altrua.features.user.domain.User;
+import com.techfun.altrua.features.user.domain.model.User;
 
 /**
  * Repositório Spring Data JPA para a entidade {@link User}.
